reorganized optionalrefactoring.core packages
