+++ edu.cuny.hunter.optionalrefactoring.core/src/edu/cuny/hunter/optionalrefactoring/core/analysis/Preconditions.java
@@ -1,4 +1,4 @@
-package edu.cuny.hunter.optionalrefactoring.core.refactorings;
+package edu.cuny.hunter.optionalrefactoring.core.analysis;
 
 import org.eclipse.jdt.core.dom.ArrayCreation;
 import org.eclipse.jdt.core.dom.ArrayInitializer;
