+++ edu.cuny.hunter.optionalrefactoring.core/src/edu/cuny/hunter/optionalrefactoring/core/utils/Visitor.java
@@ -1,5 +1,9 @@
 package edu.cuny.hunter.optionalrefactoring.core.utils;
 
+import edu.cuny.hunter.optionalrefactoring.core.refactorings.ComputationNode;
+import edu.cuny.hunter.optionalrefactoring.core.refactorings.UnionComputationNode;
+import edu.cuny.hunter.optionalrefactoring.core.refactorings.ValuedComputationNode;
+
 public interface Visitor {
 	public void visit(ComputationNode node);
