+++ edu.cuny.hunter.optionalrefactoring.tests/test cases/edu/cuny/hunter/optionalrefactoring/ui/tests/ScratchClass.java
@@ -1,41 +1,11 @@
 package edu.cuny.hunter.optionalrefactoring.ui.tests;
 
+import java.util.Optional;
+
 import org.eclipse.jdt.core.dom.Assignment;
 
 class ScratchClass {
 	
-	public class A {
-		
-		public class B {
-			// should seed: {k,l,m,n,o,h,j,d,c,a}
-			// should propagate: {{k},{l},{m},{n},{o},{h},{j},{d},{c},{a}}
-			public B() {
-				this(null);
-			}
-			
-			public B(Object a) {
-				this(new Object(), null);
-			}
-			
-			public B(Object b, Object c) {	
-				this(null, new Object(), new Object());
-			}
-			
-			public B(Object d, Object e, Object f) {	
-				this(new Object(), null, new Object(), null);
-			}
-			
-			public B(Object g, Object h, Object i, Object j) {	
-				this(null, null, null, null, null);
-			}
-
-			public B(Object k, Object l, Object m, Object n, Object o) {	}
-		}
-		
-		Assignment x;
-		
-		A a;
-		A b = a;
-		A c, d, localVarControl = new A();
-	}
+	
+	
 }
